refactor: move util functions to separate script
